Datagrid refactor
