+++ app/components/analysis/wsdot-viewer.vue
@@ -196,7 +196,7 @@
 
 <script lang="ts" setup>
 import type { ComputedRef } from 'vue'
-import { fmtDate, type Feature } from '~~/src/core'
+import { fmtDate, type Feature, type TableColumn, type TableReport } from '~~/src/core'
 import {
   SERVICE_LEVELS,
   levelColors,
@@ -206,7 +206,6 @@ import type {
   WSDOTReportConfig,
   LevelKey
 } from '~~/src/analysis/wsdot'
-import type { TableColumn, TableReport } from '~/components/cal/datagrid.vue'
 
 // Define read-only props
 const props = defineProps<{
