+++ src/core/index.ts
@@ -11,3 +11,4 @@ export * from './census-intersection'
 export * from './census-buffer'
 export * from './choropleth'
 export * from './strings'
+export * from './table'
